refresh nodes list when go back from deploy page
