+++ ui/src/ducks/app/nodes.js
@@ -17,6 +17,13 @@ import {
   addNotificationErrorAction
 } from './notifications';
 
+import {
+  isJobCompleted,
+  getJidFromNameLocalStorage,
+  updateJobLocalStorage,
+  removeJobLocalStorage
+} from '../../services/utils';
+
 // Actions
 const FETCH_NODES = 'FETCH_NODES';
 export const SET_NODES = 'SET_NODES';
@@ -27,8 +34,6 @@ const DEPLOY_NODE = 'DEPLOY_NODE';
 const CONNECT_SALT_API = 'CONNECT_SALT_API';
 const UPDATE_EVENTS = 'UPDATE_EVENTS';
 const SUBSCRIBE_DEPLOY_EVENTS = 'SUBSCRIBE_DEPLOY_EVENTS';
-
-const JOBS = 'JOBS';
 
 let eventSrc, channel;
 
@@ -112,11 +117,6 @@ export const subscribeDeployEventsAction = jid => {
 export function* fetchNodes() {
   const result = yield call(Api.getNodes);
   if (!result.error) {
-    yield all(
-      result.body.items.map(node => {
-        return call(removeCompletedJobFromLocalStorage, node.metadata.name);
-      })
-    );
     yield put(
       setNodesAction(
         result.body.items.map(node => {
@@ -142,6 +142,11 @@ export function* fetchNodes() {
           };
         })
       )
+    );
+    yield all(
+      result.body.items.map(node => {
+        return call(removeCompletedJobFromLocalStorage, node.metadata.name);
+      })
     );
   }
 }
@@ -207,50 +212,8 @@ export function* deployNode({ payload }) {
       })
     );
   } else {
-    yield call(
-      Api.lookupJid,
-      api.url_salt,
-      salt.data.return[0].token,
-      result.data.return[0].jid
-    );
     updateJobLocalStorage(result.data.return[0].jid, payload.name);
     yield call(history.push, `/nodes/deploy/${result.data.return[0].jid}`);
-  }
-}
-
-function isJobCompleted(result, jid) {
-  return (
-    result.return[0][jid] &&
-    Object.keys(result.return[0][jid]['Result']).length &&
-    Object.values(result.return[0][jid]['Result'])[0].return.success
-  );
-}
-
-function getJidFromNameLocalStorage(name) {
-  const jobs = JSON.parse(localStorage.getItem(JOBS)) || [];
-  const job = jobs.find(job => job.name === name);
-  return job && job.jid;
-}
-
-function getNameFromJidLocalStorage(jid) {
-  const jobs = JSON.parse(localStorage.getItem(JOBS)) || [];
-  const job = jobs.find(job => job.jid === jid);
-  return job && job.name;
-}
-
-function updateJobLocalStorage(jid, name) {
-  const jobs = JSON.parse(localStorage.getItem(JOBS)) || [];
-  jobs.push({ jid, name });
-  localStorage.setItem(JOBS, JSON.stringify(jobs));
-}
-
-function removeJobLocalStorage(jid) {
-  let jobs = JSON.parse(localStorage.getItem(JOBS)) || [];
-  jobs = jobs.filter(job => job.jid !== jid);
-  if (jobs.length) {
-    localStorage.setItem(JOBS, JSON.stringify(jobs));
-  } else {
-    localStorage.removeItem(JOBS);
   }
 }
